refactor(modules): create entry points for each module
- create single entry point for each module
- expose module contents via re-export
