+++ src/main.ts
@@ -6,9 +6,10 @@ import { APP_BASE_HREF, ROUTER_PROVIDERS } from 'angular2/router';
 // root component
 import { App } from 'components/app/app';
 
-// modules
-import { API_PROVIDERS } from 'modules/api/providers';
-import { TASK_PROVIDERS } from 'modules/task/providers';
+// module providers
+import { API_PROVIDERS } from 'modules/api';
+import { STORE_PROVIDERS } from 'modules/store';
+import { TASK_PROVIDERS } from 'modules/task';
 
 // global styles
 import 'styles/styles.scss';
@@ -18,6 +19,7 @@ bootstrap(App, [
   HTTP_PROVIDERS,
   ROUTER_PROVIDERS,
   API_PROVIDERS,
+  STORE_PROVIDERS,
   TASK_PROVIDERS,
   provide(APP_BASE_HREF, {useValue: '/'})
 ]).catch((error: Error) => console.error(error));
