+++ src/modules/api.ts
@@ -1,0 +1,5 @@
+import { ApiService } from './api/api-service';
+
+
+export { ApiService };
+export const API_PROVIDERS: any[] = [ApiService];
