--- src/modules/api/providers.ts
@@ -1,6 +1,0 @@
-import { ApiService } from './api-service';
-
-
-export const API_PROVIDERS: any[] = [
-  ApiService
-];
