+++ src/modules/task.ts
@@ -1,0 +1,7 @@
+import { TaskService } from './task/task-service';
+
+
+export { TaskService };
+export * from './task/task';
+export * from './task/task-reducer';
+export const TASK_PROVIDERS: any[] = [TaskService];
