--- src/modules/task/providers.ts
@@ -1,8 +1,0 @@
-import { provideStore } from '@ngrx/store';
-import { TaskService } from './task-service';
-import { INITIAL_STATE, taskReducer } from './task-reducer';
-
-
-export const TASK_PROVIDERS: any[] = [
-  TaskService
-].concat(provideStore({tasks: taskReducer}, {tasks: INITIAL_STATE}));
